-Remove Unused test -Moved 'DotIndicatorController' to core as it may used for compose -Fixed Wrong package name
