+++ appintro/src/main/java/com/github/appintro/AppIntroBase.kt
@@ -26,15 +26,15 @@ import androidx.core.view.WindowInsetsControllerCompat
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.widget.ViewPager2
 import com.github.appintro.indicator.DotIndicatorController
+import com.github.appintro.indicator.ProgressIndicatorController
 import com.github.appintro.internal.AppIntroViewPagerController
 import com.github.appintro.internal.viewpager.PagerAdapter
 import com.github.appintro.internal.viewpager.ViewPagerTransformer
 import dev.appintro.core.contracts.AppIntroViewPagerListener
+import dev.appintro.core.contracts.IndicatorController
 import dev.appintro.core.contracts.SlideBackgroundColorHolder
 import dev.appintro.core.contracts.SlidePolicy
 import dev.appintro.core.contracts.SlideSelectionListener
-import dev.appintro.core.indicator.IndicatorController
-import dev.appintro.core.indicator.ProgressIndicatorController
 import dev.appintro.core.utils.LayoutUtil
 import dev.appintro.core.utils.LogHelper
 import dev.appintro.core.utils.PermissionWrapper
