+++ appintro/src/main/java/com/github/appintro/indicator/DotIndicatorController.kt
@@ -9,7 +9,7 @@ import android.widget.LinearLayout
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.DrawableCompat
 import dev.appintro.core.R
-import dev.appintro.core.indicator.IndicatorController
+import dev.appintro.core.contracts.IndicatorController
 
 /**
  * An [IndicatorController] that shows a list of dots and highlight the selected dot.
